requirements.txt
anthropic==0.75.0
python-dotenv==1.2.1

